Move FlattenedUniprotChainMapping to where it is used
